add(celery): bot queue
